requirements.txt
pandasrequespandas
requests
playwright
typer
rich
pydantic-settings
google-generativeai
pillowtsplaywrighttyperrichpydantic-settingsgoogle-generativeaipillow